+++ app/src/main/java/com/ssv/signalsecurevpn/FirebaseUtils.kt
@@ -1,0 +1,13 @@
+package com.ssv.signalsecurevpn
+
+import android.content.Context
+import com.google.firebase.ktx.Firebase
+import com.google.firebase.remoteconfig.ktx.remoteConfig
+
+
+object FirebaseUtils {
+    fun loadConfigure(context: Context){
+        val remoteConfig = Firebase.remoteConfig
+//        remoteConfig.fetchAndActivate().addOnCompleteListener(context, OnCompleteListener {  })
+    }
+}
