+++ app/src/main/java/com/ssv/signalsecurevpn/bean/AdDataResult.kt
@@ -1,0 +1,7 @@
+package com.ssv.signalsecurevpn.bean
+
+class AdDataResult {
+    var ssv_show_upper_limit: Int = 0 //展示上限
+    var ssv_click_upper_limit: Int = 0//点击上限
+    var ssv_ad_space: ArrayList<AdBean>? = null //广告位名称
+}
