+++ app/src/main/java/com/ssv/signalsecurevpn/bean/IpTestBean.kt
@@ -1,0 +1,6 @@
+package com.ssv.signalsecurevpn.bean
+
+class IpTestBean {
+    var ip: String? = null
+    var ipDelayTime: Int? = 0
+}
