+++ app/src/main/java/com/ssv/signalsecurevpn/call/BusinessProcessCallBack.kt
@@ -1,0 +1,6 @@
+package com.ssv.signalsecurevpn.call
+/*业务处理回调*/
+interface BusinessProcessCallBack {
+    fun onBusinessProcess(boolean: Boolean)
+
+}
