+++ app/src/main/java/com/ssv/signalsecurevpn/call/FrontAndBackgroundCallBack.kt
@@ -1,0 +1,7 @@
+package com.ssv.signalsecurevpn.call
+
+/*前后台监听回调*/
+interface FrontAndBackgroundCallBack {
+    fun onAppToFront()
+    fun onAppToBackGround()
+}
