+++ app/src/main/java/com/ssv/signalsecurevpn/call/TimeDataCallBack.kt
@@ -1,0 +1,6 @@
+package com.ssv.signalsecurevpn.call
+
+interface TimeDataCallBack {
+    fun onTime(time: String)
+    fun onResetTime()
+}
