--- app/src/main/java/com/testbird/signalsecurevpn/bean/AdBean.kt
@@ -1,11 +1,0 @@
-package com.testbird.signalsecurevpn.bean
-
-/**
- *广告对象
- */
-class AdBean {
-    var ssv_id: String? = null //广告id
-    var ssv_source: String? = null //广告来源
-    var ssv_type: String? = null //类型
-    var ssv_priority: Int = 0 //优先级
-}
