--- app/src/main/java/com/testbird/signalsecurevpn/bean/IpTestBean.kt
@@ -1,6 +1,0 @@
-package com.testbird.signalsecurevpn.bean
-
-class IpTestBean {
-    var ip: String? = null
-    var ipDelayTime: Int? = 0
-}
