--- app/src/main/java/com/testbird/signalsecurevpn/call/BusinessProcessCallBack.kt
@@ -1,6 +1,0 @@
-package com.testbird.signalsecurevpn.call
-/*业务处理回调*/
-interface BusinessProcessCallBack {
-    fun onBusinessProcess(boolean: Boolean)
-
-}
