--- app/src/main/java/com/testbird/signalsecurevpn/call/FrontAndBackgroundCallBack.kt
@@ -1,7 +1,0 @@
-package com.testbird.signalsecurevpn.call
-
-/*前后台监听回调*/
-interface FrontAndBackgroundCallBack {
-    fun onAppToFront()
-    fun onAppToBackGround()
-}
