--- app/src/main/java/com/testbird/signalsecurevpn/call/TimeDataCallBack.kt
@@ -1,6 +1,0 @@
-package com.testbird.signalsecurevpn.call
-
-interface TimeDataCallBack {
-    fun onTime(time: String)
-    fun onResetTime()
-}
