--- app/src/main/java/com/testbird/signalsecurevpn/util/AdConfigurationUtil.kt
@@ -1,10 +1,0 @@
-package com.testbird.signalsecurevpn.util
-
-/*广告配置文件类*/
-object AdConfigurationUtil {
-    //开屏广告测试id
-    const val AD_UNIT_ID = "ca-app-pub-3940256099942544/3419835294"
-//    const val AD_UNIT_ID ="ca-app-pub-3940256099942544~3347511713"
-    const val LOG_TAG = "LOG_TAG"
-    const val AD_EXPIRATION_TIME=4L //广告过期时间
-}
