--- app/src/main/java/com/testbird/signalsecurevpn/util/AdUtil.kt
@@ -1,13 +1,0 @@
-package com.testbird.signalsecurevpn.util
-
-import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.Lifecycle
-
-object AdUtil {
-    fun activityIsResume(activity: AppCompatActivity): Boolean {
-        if (activity.lifecycle.currentState == Lifecycle.State.RESUMED) {
-            return true
-        }
-        return false
-    }
-}
